refactor: split docker-manager tests into focused files
Split the 5472-line docker-manager.test.ts into focused test
files: utils, compose, lifecycle, and cleanup. Each file has
its own mock setup and can be run independently.

Closes #2474
