+++ src/docker-manager-compose.test.ts
@@ -1,4 +1,4 @@
-import { generateDockerCompose, subnetsOverlap, writeConfigs, startContainers, stopContainers, fastKillAgentContainer, isAgentExternallyKilled, resetAgentExternallyKilled, AGENT_CONTAINER_NAME, cleanup, runAgentCommand, validateIdNotInSystemRange, getSafeHostUid, getSafeHostGid, getRealUserHome, extractGhHostFromServerUrl, readGitHubPathEntries, mergeGitHubPathEntries, readGitHubEnvEntries, parseGitHubEnvFile, readEnvFile, MIN_REGULAR_UID, ACT_PRESET_BASE_IMAGE, stripScheme, collectDiagnosticLogs, setAwfDockerHost } from './docker-manager';
+import { generateDockerCompose, readGitHubPathEntries, mergeGitHubPathEntries, readGitHubEnvEntries, parseGitHubEnvFile, readEnvFile, ACT_PRESET_BASE_IMAGE, AGENT_CONTAINER_NAME } from './docker-manager';
 import { WrapperConfig } from './types';
 import * as fs from 'fs';
 import * as path from 'path';
@@ -15,308 +15,7 @@ jest.mock('execa', () => {
   return fn;
 });
 
-describe('docker-manager', () => {
-  describe('subnetsOverlap', () => {
-
-    it('should detect overlapping subnets with same CIDR', () => {
-      expect(subnetsOverlap('172.30.0.0/24', '172.30.0.0/24')).toBe(true);
-    });
-
-    it('should detect non-overlapping subnets', () => {
-      expect(subnetsOverlap('172.30.0.0/24', '172.31.0.0/24')).toBe(false);
-      expect(subnetsOverlap('192.168.1.0/24', '192.168.2.0/24')).toBe(false);
-    });
-
-    it('should detect when smaller subnet is inside larger subnet', () => {
-      expect(subnetsOverlap('172.16.0.0/16', '172.16.5.0/24')).toBe(true);
-      expect(subnetsOverlap('172.16.5.0/24', '172.16.0.0/16')).toBe(true);
-    });
-
-    it('should detect partial overlap', () => {
-      expect(subnetsOverlap('172.30.0.0/23', '172.30.1.0/24')).toBe(true);
-    });
-
-    it('should handle Docker default bridge network', () => {
-      expect(subnetsOverlap('172.17.0.0/16', '172.17.5.0/24')).toBe(true);
-      expect(subnetsOverlap('172.17.0.0/16', '172.18.0.0/16')).toBe(false);
-    });
-
-    it('should handle /32 (single host) networks', () => {
-      expect(subnetsOverlap('192.168.1.1/32', '192.168.1.1/32')).toBe(true);
-      expect(subnetsOverlap('192.168.1.1/32', '192.168.1.2/32')).toBe(false);
-    });
-  });
-
-  describe('ACT_PRESET_BASE_IMAGE', () => {
-    it('should be a valid catthehacker act image', () => {
-      expect(ACT_PRESET_BASE_IMAGE).toBe('ghcr.io/catthehacker/ubuntu:act-24.04');
-    });
-
-    it('should match expected pattern for catthehacker images', () => {
-      expect(ACT_PRESET_BASE_IMAGE).toMatch(/^ghcr\.io\/catthehacker\/ubuntu:act-\d+\.\d+$/);
-    });
-  });
-
-  describe('validateIdNotInSystemRange', () => {
-    it('should return 1000 for system UIDs (0-999)', () => {
-      expect(validateIdNotInSystemRange(0)).toBe('1000');
-      expect(validateIdNotInSystemRange(1)).toBe('1000');
-      expect(validateIdNotInSystemRange(13)).toBe('1000'); // proxy user
-      expect(validateIdNotInSystemRange(999)).toBe('1000');
-    });
-
-    it('should return the UID as-is for regular users (>= 1000)', () => {
-      expect(validateIdNotInSystemRange(1000)).toBe('1000');
-      expect(validateIdNotInSystemRange(1001)).toBe('1001');
-      expect(validateIdNotInSystemRange(65534)).toBe('65534'); // nobody user on some systems
-    });
-  });
-
-  describe('getSafeHostUid', () => {
-    const originalGetuid = process.getuid;
-    const originalSudoUid = process.env.SUDO_UID;
-
-    afterEach(() => {
-      process.getuid = originalGetuid;
-      if (originalSudoUid !== undefined) {
-        process.env.SUDO_UID = originalSudoUid;
-      } else {
-        delete process.env.SUDO_UID;
-      }
-    });
-
-    it('should return 1000 when SUDO_UID is a system UID', () => {
-      process.getuid = () => 0; // Running as root
-      process.env.SUDO_UID = '13'; // proxy user
-      expect(getSafeHostUid()).toBe('1000');
-    });
-
-    it('should return SUDO_UID when it is a regular user UID', () => {
-      process.getuid = () => 0; // Running as root
-      process.env.SUDO_UID = '1001';
-      expect(getSafeHostUid()).toBe('1001');
-    });
-
-    it('should return 1000 when SUDO_UID is 0', () => {
-      process.getuid = () => 0; // Running as root
-      process.env.SUDO_UID = '0';
-      expect(getSafeHostUid()).toBe('1000');
-    });
-
-    it('should return 1000 when running as root without SUDO_UID', () => {
-      process.getuid = () => 0;
-      delete process.env.SUDO_UID;
-      expect(getSafeHostUid()).toBe('1000');
-    });
-
-    it('should return 1000 for non-root system UID', () => {
-      process.getuid = () => 13; // proxy user
-      delete process.env.SUDO_UID;
-      expect(getSafeHostUid()).toBe('1000');
-    });
-
-    it('should return the UID when running as regular user', () => {
-      process.getuid = () => 1001;
-      delete process.env.SUDO_UID;
-      expect(getSafeHostUid()).toBe('1001');
-    });
-
-    it('should return 1000 when SUDO_UID is not a valid number', () => {
-      process.getuid = () => 0; // Running as root
-      process.env.SUDO_UID = 'not-a-number';
-      expect(getSafeHostUid()).toBe('1000');
-    });
-  });
-
-  describe('getSafeHostGid', () => {
-    const originalGetgid = process.getgid;
-    const originalSudoGid = process.env.SUDO_GID;
-
-    afterEach(() => {
-      process.getgid = originalGetgid;
-      if (originalSudoGid !== undefined) {
-        process.env.SUDO_GID = originalSudoGid;
-      } else {
-        delete process.env.SUDO_GID;
-      }
-    });
-
-    it('should return 1000 when SUDO_GID is a system GID', () => {
-      process.getgid = () => 0; // Running as root
-      process.env.SUDO_GID = '13'; // proxy group
-      expect(getSafeHostGid()).toBe('1000');
-    });
-
-    it('should return SUDO_GID when it is a regular user GID', () => {
-      process.getgid = () => 0; // Running as root
-      process.env.SUDO_GID = '1001';
-      expect(getSafeHostGid()).toBe('1001');
-    });
-
-    it('should return 1000 when SUDO_GID is 0', () => {
-      process.getgid = () => 0; // Running as root
-      process.env.SUDO_GID = '0';
-      expect(getSafeHostGid()).toBe('1000');
-    });
-
-    it('should return 1000 when running as root without SUDO_GID', () => {
-      process.getgid = () => 0;
-      delete process.env.SUDO_GID;
-      expect(getSafeHostGid()).toBe('1000');
-    });
-
-    it('should return 1000 for non-root system GID', () => {
-      process.getgid = () => 13; // proxy group
-      delete process.env.SUDO_GID;
-      expect(getSafeHostGid()).toBe('1000');
-    });
-
-    it('should return the GID when running as regular user', () => {
-      process.getgid = () => 1001;
-      delete process.env.SUDO_GID;
-      expect(getSafeHostGid()).toBe('1001');
-    });
-
-    it('should return 1000 when SUDO_GID is not a valid number', () => {
-      process.getgid = () => 0; // Running as root
-      process.env.SUDO_GID = 'not-a-number';
-      expect(getSafeHostGid()).toBe('1000');
-    });
-  });
-
-  describe('getRealUserHome', () => {
-    const originalGetuid = process.getuid;
-    const originalSudoUser = process.env.SUDO_USER;
-    const originalHome = process.env.HOME;
-
-    afterEach(() => {
-      process.getuid = originalGetuid;
-      process.env.SUDO_USER = originalSudoUser;
-      process.env.HOME = originalHome;
-      jest.restoreAllMocks();
-    });
-
-    it('should return HOME when running as regular user', () => {
-      process.getuid = () => 1001;
-      process.env.HOME = '/home/testuser';
-      expect(getRealUserHome()).toBe('/home/testuser');
-    });
-
-    it('should return /root as fallback when HOME is not set and running as root', () => {
-      process.getuid = () => 0;
-      delete process.env.SUDO_USER;
-      delete process.env.HOME;
-      expect(getRealUserHome()).toBe('/root');
-    });
-
-    it('should use HOME as fallback when running as root without SUDO_USER', () => {
-      process.getuid = () => 0;
-      delete process.env.SUDO_USER;
-      process.env.HOME = '/root';
-      expect(getRealUserHome()).toBe('/root');
-    });
-
-    it('should look up user home from /etc/passwd when running as root with SUDO_USER (using real root user)', () => {
-      // Test with actual /etc/passwd by using 'root' user which always exists
-      process.getuid = () => 0;
-      process.env.SUDO_USER = 'root';
-      process.env.HOME = '/some/other/path';
-
-      // Should find root's home directory from /etc/passwd
-      expect(getRealUserHome()).toBe('/root');
-    });
-
-    it('should fall back to HOME when SUDO_USER not found in /etc/passwd', () => {
-      process.getuid = () => 0;
-      process.env.SUDO_USER = 'nonexistent_user_12345';
-      process.env.HOME = '/fallback/home';
-
-      // User doesn't exist in /etc/passwd, should fall back to HOME
-      expect(getRealUserHome()).toBe('/fallback/home');
-    });
-
-    it('should handle undefined getuid gracefully (using real /etc/passwd)', () => {
-      // Simulate environment where process.getuid is undefined (e.g., Windows)
-      process.getuid = undefined as any;
-      process.env.SUDO_USER = 'root';
-      process.env.HOME = '/custom/home';
-
-      // With getuid undefined, uid is undefined (falsy), so it attempts passwd lookup
-      // Should find root's home directory from /etc/passwd
-      expect(getRealUserHome()).toBe('/root');
-    });
-  });
-
-  describe('extractGhHostFromServerUrl', () => {
-    it('should return null for undefined GITHUB_SERVER_URL', () => {
-      expect(extractGhHostFromServerUrl(undefined)).toBeNull();
-    });
-
-    it('should return null for empty string GITHUB_SERVER_URL', () => {
-      expect(extractGhHostFromServerUrl('')).toBeNull();
-    });
-
-    it('should return null for github.com (public GitHub)', () => {
-      expect(extractGhHostFromServerUrl('https://github.com')).toBeNull();
-    });
-
-    it('should extract hostname for GHEC instance (*.ghe.com)', () => {
-      expect(extractGhHostFromServerUrl('https://acme.ghe.com')).toBe('acme.ghe.com');
-    });
-
-    it('should extract hostname for GHES instance', () => {
-      expect(extractGhHostFromServerUrl('https://github.company.com')).toBe('github.company.com');
-    });
-
-    it('should extract hostname for GHES instance with custom port', () => {
-      expect(extractGhHostFromServerUrl('https://github.internal:8443')).toBe('github.internal');
-    });
-
-    it('should handle GITHUB_SERVER_URL without trailing slash', () => {
-      expect(extractGhHostFromServerUrl('https://github.enterprise.local')).toBe('github.enterprise.local');
-    });
-
-    it('should handle GITHUB_SERVER_URL with trailing slash', () => {
-      expect(extractGhHostFromServerUrl('https://github.enterprise.local/')).toBe('github.enterprise.local');
-    });
-
-    it('should return null for invalid URL', () => {
-      expect(extractGhHostFromServerUrl('not-a-valid-url')).toBeNull();
-    });
-
-    it('should return null for malformed URL', () => {
-      expect(extractGhHostFromServerUrl('http://')).toBeNull();
-    });
-  });
-
-  describe('MIN_REGULAR_UID constant', () => {
-    it('should be 1000 (standard Linux regular user UID threshold)', () => {
-      expect(MIN_REGULAR_UID).toBe(1000);
-    });
-  });
-
-  describe('stripScheme', () => {
-    it('should strip https:// prefix', () => {
-      expect(stripScheme('https://my-gateway.example.com')).toBe('my-gateway.example.com');
-    });
-
-    it('should strip http:// prefix', () => {
-      expect(stripScheme('http://my-gateway.example.com')).toBe('my-gateway.example.com');
-    });
-
-    it('should preserve bare hostname', () => {
-      expect(stripScheme('api.openai.com')).toBe('api.openai.com');
-    });
-
-    it('should normalize URL with path to hostname only', () => {
-      expect(stripScheme('https://my-gateway.example.com/some-path')).toBe('my-gateway.example.com');
-    });
-
-    it('should not strip scheme-like substrings in the middle', () => {
-      expect(stripScheme('api.https.example.com')).toBe('api.https.example.com');
-    });
-  });
-
+describe('docker-manager generateDockerCompose', () => {
   describe('generateDockerCompose', () => {
     const mockConfig: WrapperConfig = {
       allowedDomains: ['github.com', 'npmjs.org'],
@@ -3631,1339 +3330,6 @@ describe('docker-manager', () => {
       });
     });
   });
-
-  describe('writeConfigs', () => {
-    let testDir: string;
-
-    beforeEach(() => {
-      testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-test-'));
-      jest.clearAllMocks();
-    });
-
-    afterEach(() => {
-      if (fs.existsSync(testDir)) {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should create work directory if it does not exist', async () => {
-      const newWorkDir = path.join(testDir, 'new-work-dir');
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: newWorkDir,
-      };
-
-      // writeConfigs may succeed if seccomp profile is found, or fail if not
-      try {
-        await writeConfigs(config);
-      } catch {
-        // Expected to fail if seccomp profile not found, but directories should still be created
-      }
-
-      // Verify work directory was created
-      expect(fs.existsSync(newWorkDir)).toBe(true);
-    });
-
-    it('should create agent-logs directory', async () => {
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail, but directories should still be created
-      }
-
-      // Verify agent-logs directory was created
-      expect(fs.existsSync(path.join(testDir, 'agent-logs'))).toBe(true);
-    });
-
-    it('should create squid-logs directory', async () => {
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail, but directories should still be created
-      }
-
-      // Verify squid-logs directory was created
-      expect(fs.existsSync(path.join(testDir, 'squid-logs'))).toBe(true);
-    });
-
-    it('should create /tmp/gh-aw/mcp-logs directory with world-writable permissions', async () => {
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail, but directory should still be created
-      }
-
-      // Verify /tmp/gh-aw/mcp-logs directory was created
-      expect(fs.existsSync('/tmp/gh-aw/mcp-logs')).toBe(true);
-      const stats = fs.statSync('/tmp/gh-aw/mcp-logs');
-      expect(stats.isDirectory()).toBe(true);
-      // Verify permissions are 0o777 (rwxrwxrwx) to allow non-root users to create subdirectories
-      expect((stats.mode & 0o777).toString(8)).toBe('777');
-    });
-
-    it('should write squid.conf file', async () => {
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com', 'example.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail after writing configs
-      }
-
-      // Verify squid.conf was created (it's created before seccomp check)
-      const squidConfPath = path.join(testDir, 'squid.conf');
-      if (fs.existsSync(squidConfPath)) {
-        const content = fs.readFileSync(squidConfPath, 'utf-8');
-        expect(content).toContain('github.com');
-        expect(content).toContain('example.com');
-      }
-    });
-
-    it('should write docker-compose.yml file', async () => {
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail after writing configs
-      }
-
-      // Verify docker-compose.yml was created
-      const dockerComposePath = path.join(testDir, 'docker-compose.yml');
-      if (fs.existsSync(dockerComposePath)) {
-        const content = fs.readFileSync(dockerComposePath, 'utf-8');
-        expect(content).toContain('awf-squid');
-        expect(content).toContain('awf-agent');
-      }
-    });
-
-    it('should create work directory with restricted permissions (0o700)', async () => {
-      const newWorkDir = path.join(testDir, 'restricted-dir');
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: newWorkDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail if seccomp profile not found
-      }
-
-      // Verify directory was created with restricted permissions
-      expect(fs.existsSync(newWorkDir)).toBe(true);
-      const stats = fs.statSync(newWorkDir);
-      expect((stats.mode & 0o777).toString(8)).toBe('700');
-    });
-
-    it('should write config files with restricted permissions (0o600)', async () => {
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail after writing configs
-      }
-
-      // Verify squid.conf is readable by proxy user (0o644) for non-root Squid
-      const squidConfPath = path.join(testDir, 'squid.conf');
-      if (fs.existsSync(squidConfPath)) {
-        const stats = fs.statSync(squidConfPath);
-        expect((stats.mode & 0o777).toString(8)).toBe('644');
-      }
-
-      // Verify docker-compose.yml has restricted permissions
-      const dockerComposePath = path.join(testDir, 'docker-compose.yml');
-      if (fs.existsSync(dockerComposePath)) {
-        const stats = fs.statSync(dockerComposePath);
-        expect((stats.mode & 0o777).toString(8)).toBe('600');
-      }
-    });
-
-    it('should use proxyLogsDir when specified', async () => {
-      const proxyLogsDir = path.join(testDir, 'custom-proxy-logs');
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-        proxyLogsDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail after writing configs
-      }
-
-      // Verify proxyLogsDir was created
-      expect(fs.existsSync(proxyLogsDir)).toBe(true);
-    });
-
-    it('should create api-proxy-logs subdirectory inside proxyLogsDir when specified', async () => {
-      const proxyLogsDir = path.join(testDir, 'custom-proxy-logs');
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-        proxyLogsDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail after writing configs
-      }
-
-      // Verify api-proxy-logs subdirectory was created inside proxyLogsDir
-      const apiProxyLogsDir = path.join(proxyLogsDir, 'api-proxy-logs');
-      expect(fs.existsSync(apiProxyLogsDir)).toBe(true);
-    });
-
-    it('should create proxyLogsDir with nested non-existent parents', async () => {
-      const proxyLogsDir = path.join(testDir, 'deeply', 'nested', 'proxy-logs');
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-        proxyLogsDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail after writing configs
-      }
-
-      // Verify deeply nested proxyLogsDir was created recursively
-      expect(fs.existsSync(proxyLogsDir)).toBe(true);
-    });
-
-    it('should pre-create chroot home subdirectories with correct ownership', async () => {
-      // Use a temporary home directory to avoid modifying the real one
-      const fakeHome = path.join(testDir, 'fakehome');
-      fs.mkdirSync(fakeHome, { recursive: true });
-      const originalHome = process.env.HOME;
-      const originalSudoUser = process.env.SUDO_USER;
-      process.env.HOME = fakeHome;
-      // Clear SUDO_USER to make getRealUserHome() use process.env.HOME
-      delete process.env.SUDO_USER;
-
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail after writing configs
-      }
-
-      // Verify chroot home subdirectories were created
-      const expectedDirs = [
-        '.copilot', '.cache', '.config', '.local',
-        '.anthropic', '.claude', '.cargo', '.rustup', '.npm', '.nvm',
-      ];
-      for (const dir of expectedDirs) {
-        expect(fs.existsSync(path.join(fakeHome, dir))).toBe(true);
-      }
-      // ~/.gemini is only pre-created when geminiApiKey is configured
-      expect(fs.existsSync(path.join(fakeHome, '.gemini'))).toBe(false);
-
-      if (originalHome !== undefined) {
-        process.env.HOME = originalHome;
-      } else {
-        delete process.env.HOME;
-      }
-      if (originalSudoUser !== undefined) {
-        process.env.SUDO_USER = originalSudoUser;
-      } else {
-        delete process.env.SUDO_USER;
-      }
-    });
-
-    it('should pre-create ~/.gemini when geminiApiKey is configured', async () => {
-      const fakeHome = path.join(testDir, 'fakehome-gemini');
-      fs.mkdirSync(fakeHome, { recursive: true });
-      const originalHome = process.env.HOME;
-      const originalSudoUser = process.env.SUDO_USER;
-      process.env.HOME = fakeHome;
-      delete process.env.SUDO_USER;
-
-      const config: WrapperConfig = {
-        allowedDomains: ['github.com'],
-        agentCommand: 'echo test',
-        logLevel: 'info',
-        keepContainers: false,
-        workDir: testDir,
-        geminiApiKey: 'AIza-test-key',
-      };
-
-      try {
-        await writeConfigs(config);
-      } catch {
-        // May fail after writing configs
-      }
-
-      expect(fs.existsSync(path.join(fakeHome, '.gemini'))).toBe(true);
-
-      if (originalHome !== undefined) {
-        process.env.HOME = originalHome;
-      } else {
-        delete process.env.HOME;
-      }
-      if (originalSudoUser !== undefined) {
-        process.env.SUDO_USER = originalSudoUser;
-      } else {
-        delete process.env.SUDO_USER;
-      }
-    });
-  });
-
-  describe('startContainers', () => {
-    let testDir: string;
-
-    beforeEach(() => {
-      testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-test-'));
-      jest.clearAllMocks();
-    });
-
-    afterEach(() => {
-      if (fs.existsSync(testDir)) {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should remove existing containers before starting', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await startContainers(testDir, ['github.com']);
-
-      expect(mockExecaFn).toHaveBeenCalledWith(
-        'docker',
-        ['rm', '-f', 'awf-squid', 'awf-agent', 'awf-iptables-init', 'awf-api-proxy', 'awf-cli-proxy'],
-        expect.objectContaining({ reject: false })
-      );
-    });
-
-    it('should continue when removing existing containers fails', async () => {
-      // First call (docker rm) throws an error, but we should continue
-      mockExecaFn.mockRejectedValueOnce(new Error('No such container'));
-      // Second call (docker compose up) succeeds
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await startContainers(testDir, ['github.com']);
-
-      // Should still call docker compose up even if rm failed
-      expect(mockExecaFn).toHaveBeenCalledWith(
-        'docker',
-        ['compose', 'up', '-d'],
-        expect.objectContaining({ cwd: testDir, stdout: process.stderr, stderr: 'inherit' })
-      );
-    });
-
-    it('should run docker compose up', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await startContainers(testDir, ['github.com']);
-
-      expect(mockExecaFn).toHaveBeenCalledWith(
-        'docker',
-        ['compose', 'up', '-d'],
-        expect.objectContaining({ cwd: testDir, stdout: process.stderr, stderr: 'inherit' })
-      );
-    });
-
-    it('should run docker compose up with --pull never when skipPull is true', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await startContainers(testDir, ['github.com'], undefined, true);
-
-      expect(mockExecaFn).toHaveBeenCalledWith(
-        'docker',
-        ['compose', 'up', '-d', '--pull', 'never'],
-        expect.objectContaining({ cwd: testDir, stdout: process.stderr, stderr: 'inherit' })
-      );
-    });
-
-    it('should run docker compose up without --pull never when skipPull is false', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await startContainers(testDir, ['github.com'], undefined, false);
-
-      expect(mockExecaFn).toHaveBeenCalledWith(
-        'docker',
-        ['compose', 'up', '-d'],
-        expect.objectContaining({ cwd: testDir, stdout: process.stderr, stderr: 'inherit' })
-      );
-    });
-
-    it('should handle docker compose failure', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      mockExecaFn.mockRejectedValueOnce(new Error('Docker compose failed'));
-
-      await expect(startContainers(testDir, ['github.com'])).rejects.toThrow('Docker compose failed');
-    });
-
-    it('should handle healthcheck failure with blocked domains', async () => {
-      // Create access.log with denied entries
-      const squidLogsDir = path.join(testDir, 'squid-logs');
-      fs.mkdirSync(squidLogsDir, { recursive: true });
-      fs.writeFileSync(
-        path.join(squidLogsDir, 'access.log'),
-        '1760994429.358 172.30.0.20:36274 blocked.com:443 -:- 1.1 CONNECT 403 TCP_DENIED:HIER_NONE blocked.com:443 "curl/7.81.0"\n'
-      );
-
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      mockExecaFn.mockRejectedValueOnce(new Error('is unhealthy'));
-
-      await expect(startContainers(testDir, ['github.com'])).rejects.toThrow();
-    });
-
-    it('should retry once when awf-api-proxy fails its health check', async () => {
-      // 1. docker rm (initial cleanup)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 2. docker compose up (first attempt - fails with api-proxy unhealthy)
-      mockExecaFn.mockRejectedValueOnce(new Error('dependency failed to start: container awf-api-proxy is unhealthy'));
-      // 3. docker logs --tail 50 awf-api-proxy (get logs for diagnosis)
-      mockExecaFn.mockResolvedValueOnce({ stdout: 'api-proxy startup logs', stderr: '', exitCode: 0 } as any);
-      // 4. docker compose down (cleanup before retry)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 5. docker compose up (retry - succeeds)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await expect(startContainers(testDir, ['github.com'])).resolves.toBeUndefined();
-
-      // Verify retry happened: compose up was called twice
-      const upCalls = mockExecaFn.mock.calls.filter((call: any[]) =>
-        call[0] === 'docker' && Array.isArray(call[1]) && call[1].includes('up')
-      );
-      expect(upCalls).toHaveLength(2);
-    });
-
-    it('should throw clear error when awf-api-proxy fails its health check on both attempts', async () => {
-      // 1. docker rm (initial cleanup)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 2. docker compose up (first attempt - fails)
-      mockExecaFn.mockRejectedValueOnce(new Error('dependency failed to start: container awf-api-proxy is unhealthy'));
-      // 3. docker logs (first diagnosis)
-      mockExecaFn.mockResolvedValueOnce({ stdout: 'api-proxy logs', stderr: '', exitCode: 0 } as any);
-      // 4. docker compose down (cleanup before retry)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 5. docker compose up (retry - also fails)
-      mockExecaFn.mockRejectedValueOnce(new Error('dependency failed to start: container awf-api-proxy is unhealthy'));
-      // 6. docker logs (second diagnosis)
-      mockExecaFn.mockResolvedValueOnce({ stdout: 'api-proxy logs', stderr: '', exitCode: 0 } as any);
-
-      await expect(startContainers(testDir, ['github.com'])).rejects.toThrow(
-        'AWF firewall failed to start: awf-api-proxy failed its health check on both attempts'
-      );
-    });
-
-    it('should not retry for non-api-proxy healthcheck failures', async () => {
-      // 1. docker rm (initial cleanup)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 2. docker compose up (fails with squid unhealthy - not api-proxy)
-      mockExecaFn.mockRejectedValueOnce(new Error('dependency failed to start: container awf-squid is unhealthy'));
-
-      await expect(startContainers(testDir, ['github.com'])).rejects.toThrow();
-
-      // Only one compose up call (no retry)
-      const upCalls = mockExecaFn.mock.calls.filter((call: any[]) =>
-        call[0] === 'docker' && Array.isArray(call[1]) && call[1].includes('up')
-      );
-      expect(upCalls).toHaveLength(1);
-    });
-
-    it('should route retry error through Squid diagnostics when retry fails with non-api-proxy error', async () => {
-      // Create access.log with denied entries so Squid diagnostics fire
-      const squidLogsDir = path.join(testDir, 'squid-logs');
-      fs.mkdirSync(squidLogsDir, { recursive: true });
-      fs.writeFileSync(
-        path.join(squidLogsDir, 'access.log'),
-        '1760994429.358 172.30.0.20:36274 blocked.com:443 -:- 1.1 CONNECT 403 TCP_DENIED:HIER_NONE blocked.com:443 "curl/7.81.0"\n'
-      );
-
-      // 1. docker rm (initial cleanup)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 2. docker compose up (first attempt - fails with api-proxy unhealthy)
-      mockExecaFn.mockRejectedValueOnce(new Error('dependency failed to start: container awf-api-proxy is unhealthy'));
-      // 3. docker logs (diagnosis before retry)
-      mockExecaFn.mockResolvedValueOnce({ stdout: 'some logs', stderr: '', exitCode: 0 } as any);
-      // 4. docker compose down (cleanup before retry)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 5. docker compose up (retry - fails with a different, non-api-proxy error)
-      mockExecaFn.mockRejectedValueOnce(new Error('dependency failed to start: container awf-squid is unhealthy'));
-
-      // Should surface the Squid blocked-domain error, not a raw throw
-      await expect(startContainers(testDir, ['github.com'])).rejects.toThrow('Firewall blocked access to:');
-    });
-
-    it('should not emit container logs when docker logs exits non-zero (container not found)', async () => {
-      // 1. docker rm (initial cleanup)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 2. docker compose up (fails with api-proxy unhealthy)
-      mockExecaFn.mockRejectedValueOnce(new Error('dependency failed to start: container awf-api-proxy is unhealthy'));
-      // 3. docker logs returns non-zero (container not found)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: 'No such container: awf-api-proxy', exitCode: 1 } as any);
-      // 4. docker compose down (cleanup before retry)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // 5. docker compose up (retry - succeeds)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      // Should succeed without emitting "No such container" noise at error level
-      await expect(startContainers(testDir, ['github.com'])).resolves.toBeUndefined();
-    });
-  });
-
-  describe('stopContainers', () => {
-    let testDir: string;
-
-    beforeEach(() => {
-      testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-test-'));
-      jest.clearAllMocks();
-    });
-
-    afterEach(() => {
-      if (fs.existsSync(testDir)) {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should skip stopping when keepContainers is true', async () => {
-      await stopContainers(testDir, true);
-
-      expect(mockExecaFn).not.toHaveBeenCalled();
-    });
-
-    it('should run docker compose down when keepContainers is false', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await stopContainers(testDir, false);
-
-      expect(mockExecaFn).toHaveBeenCalledWith(
-        'docker',
-        ['compose', 'down', '-v', '-t', '1'],
-        expect.objectContaining({ cwd: testDir, stdout: process.stderr, stderr: 'inherit' })
-      );
-    });
-
-    it('should throw error when docker compose down fails', async () => {
-      mockExecaFn.mockRejectedValueOnce(new Error('Docker compose down failed'));
-
-      await expect(stopContainers(testDir, false)).rejects.toThrow('Docker compose down failed');
-    });
-  });
-
-  describe('fastKillAgentContainer', () => {
-    beforeEach(() => {
-      jest.clearAllMocks();
-      resetAgentExternallyKilled();
-    });
-
-    it('should call docker stop with default 3-second timeout', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await fastKillAgentContainer();
-
-      expect(mockExecaFn).toHaveBeenCalledWith(
-        'docker',
-        ['stop', '-t', '3', AGENT_CONTAINER_NAME],
-        expect.objectContaining({ reject: false, timeout: 8000 })
-      );
-    });
-
-    it('should accept a custom stop timeout', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      await fastKillAgentContainer(5);
-
-      expect(mockExecaFn).toHaveBeenCalledWith(
-        'docker',
-        ['stop', '-t', '5', AGENT_CONTAINER_NAME],
-        expect.objectContaining({ reject: false, timeout: 10000 })
-      );
-    });
-
-    it('should not throw when docker stop fails', async () => {
-      mockExecaFn.mockRejectedValueOnce(new Error('docker not found'));
-
-      await expect(fastKillAgentContainer()).resolves.toBeUndefined();
-    });
-
-    it('should set the externally-killed flag', async () => {
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      expect(isAgentExternallyKilled()).toBe(false);
-      await fastKillAgentContainer();
-      expect(isAgentExternallyKilled()).toBe(true);
-    });
-
-    it('should set the externally-killed flag even when docker stop fails', async () => {
-      mockExecaFn.mockRejectedValueOnce(new Error('docker not found'));
-
-      await fastKillAgentContainer();
-      expect(isAgentExternallyKilled()).toBe(true);
-    });
-  });
-
-  describe('setAwfDockerHost / getLocalDockerEnv (DOCKER_HOST isolation)', () => {
-    const originalDockerHost = process.env.DOCKER_HOST;
-
-    afterEach(() => {
-      // Restore env and reset override after each test
-      if (originalDockerHost === undefined) {
-        delete process.env.DOCKER_HOST;
-      } else {
-        process.env.DOCKER_HOST = originalDockerHost;
-      }
-      setAwfDockerHost(undefined);
-      jest.clearAllMocks();
-    });
-
-    it('docker compose up should NOT forward a TCP DOCKER_HOST to the docker CLI', async () => {
-      process.env.DOCKER_HOST = 'tcp://localhost:2375';
-      const testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-test-'));
-      try {
-        mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any); // docker rm
-        mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any); // docker compose up
-
-        await startContainers(testDir, ['github.com']);
-
-        const composeCalls = mockExecaFn.mock.calls.filter(
-          (call: any[]) => call[1]?.[0] === 'compose'
-        );
-        expect(composeCalls.length).toBeGreaterThan(0);
-        const composeEnv = composeCalls[0][2]?.env as Record<string, string | undefined> | undefined;
-        // DOCKER_HOST must be absent from the env passed to docker compose
-        expect(composeEnv).toBeDefined();
-        expect(composeEnv!.DOCKER_HOST).toBeUndefined();
-      } finally {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-    });
-
-    it('docker compose up should keep a unix:// DOCKER_HOST in the env', async () => {
-      process.env.DOCKER_HOST = 'unix:///var/run/docker.sock';
-      const testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-test-'));
-      try {
-        mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any); // docker rm
-        mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any); // docker compose up
-
-        await startContainers(testDir, ['github.com']);
-
-        const composeCalls = mockExecaFn.mock.calls.filter(
-          (call: any[]) => call[1]?.[0] === 'compose'
-        );
-        expect(composeCalls.length).toBeGreaterThan(0);
-        const composeEnv = composeCalls[0][2]?.env as Record<string, string | undefined> | undefined;
-        expect(composeEnv).toBeDefined();
-        expect(composeEnv!.DOCKER_HOST).toBe('unix:///var/run/docker.sock');
-      } finally {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-    });
-
-    it('setAwfDockerHost should override DOCKER_HOST for AWF operations', async () => {
-      process.env.DOCKER_HOST = 'tcp://localhost:2375'; // Would normally be cleared
-      setAwfDockerHost('unix:///run/user/1000/docker.sock');
-      const testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-test-'));
-      try {
-        mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any); // docker rm
-        mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any); // docker compose up
-
-        await startContainers(testDir, ['github.com']);
-
-        const composeCalls = mockExecaFn.mock.calls.filter(
-          (call: any[]) => call[1]?.[0] === 'compose'
-        );
-        expect(composeCalls.length).toBeGreaterThan(0);
-        const composeEnv = composeCalls[0][2]?.env as Record<string, string | undefined> | undefined;
-        expect(composeEnv).toBeDefined();
-        expect(composeEnv!.DOCKER_HOST).toBe('unix:///run/user/1000/docker.sock');
-      } finally {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-    });
-  });
-
-  describe('runAgentCommand', () => {
-    let testDir: string;
-
-    beforeEach(() => {
-      testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-test-'));
-      jest.clearAllMocks();
-      resetAgentExternallyKilled();
-    });
-
-    afterEach(() => {
-      if (fs.existsSync(testDir)) {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should return exit code from container', async () => {
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait
-      mockExecaFn.mockResolvedValueOnce({ stdout: '0', stderr: '', exitCode: 0 } as any);
-
-      const result = await runAgentCommand(testDir, ['github.com']);
-
-      expect(result.exitCode).toBe(0);
-    });
-
-    it('should return non-zero exit code when command fails', async () => {
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait with non-zero exit code
-      mockExecaFn.mockResolvedValueOnce({ stdout: '1', stderr: '', exitCode: 0 } as any);
-
-      const result = await runAgentCommand(testDir, ['github.com']);
-
-      expect(result.exitCode).toBe(1);
-    });
-
-    it('should detect blocked domains from access log', async () => {
-      // Create access.log with denied entries
-      const squidLogsDir = path.join(testDir, 'squid-logs');
-      fs.mkdirSync(squidLogsDir, { recursive: true });
-      fs.writeFileSync(
-        path.join(squidLogsDir, 'access.log'),
-        '1760994429.358 172.30.0.20:36274 blocked.com:443 -:- 1.1 CONNECT 403 TCP_DENIED:HIER_NONE blocked.com:443 "curl/7.81.0"\n'
-      );
-
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait with non-zero exit code (command failed)
-      mockExecaFn.mockResolvedValueOnce({ stdout: '1', stderr: '', exitCode: 0 } as any);
-
-      const result = await runAgentCommand(testDir, ['github.com']);
-
-      expect(result.exitCode).toBe(1);
-      expect(result.blockedDomains).toContain('blocked.com');
-    });
-
-    it('should use proxyLogsDir when specified', async () => {
-      const proxyLogsDir = path.join(testDir, 'custom-logs');
-      fs.mkdirSync(proxyLogsDir, { recursive: true });
-      fs.writeFileSync(
-        path.join(proxyLogsDir, 'access.log'),
-        '1760994429.358 172.30.0.20:36274 blocked.com:443 -:- 1.1 CONNECT 403 TCP_DENIED:HIER_NONE blocked.com:443 "curl/7.81.0"\n'
-      );
-
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait
-      mockExecaFn.mockResolvedValueOnce({ stdout: '1', stderr: '', exitCode: 0 } as any);
-
-      const result = await runAgentCommand(testDir, ['github.com'], proxyLogsDir);
-
-      expect(result.blockedDomains).toContain('blocked.com');
-    });
-
-    it('should throw error when docker wait fails', async () => {
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait failure
-      mockExecaFn.mockRejectedValueOnce(new Error('Container not found'));
-
-      await expect(runAgentCommand(testDir, ['github.com'])).rejects.toThrow('Container not found');
-    });
-
-    it('should handle blocked domain without port (standard port 443)', async () => {
-      const squidLogsDir = path.join(testDir, 'squid-logs');
-      fs.mkdirSync(squidLogsDir, { recursive: true });
-      fs.writeFileSync(
-        path.join(squidLogsDir, 'access.log'),
-        '1760994429.358 172.30.0.20:36274 example.com:443 -:- 1.1 CONNECT 403 TCP_DENIED:HIER_NONE example.com:443 "curl/7.81.0"\n'
-      );
-
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait with non-zero exit code
-      mockExecaFn.mockResolvedValueOnce({ stdout: '1', stderr: '', exitCode: 0 } as any);
-
-      const result = await runAgentCommand(testDir, ['github.com']);
-
-      expect(result.exitCode).toBe(1);
-      expect(result.blockedDomains).toContain('example.com');
-    });
-
-    it('should handle allowed domain in blocklist correctly', async () => {
-      const squidLogsDir = path.join(testDir, 'squid-logs');
-      fs.mkdirSync(squidLogsDir, { recursive: true });
-      // Create a log entry for subdomain of allowed domain
-      fs.writeFileSync(
-        path.join(squidLogsDir, 'access.log'),
-        '1760994429.358 172.30.0.20:36274 api.github.com:8443 -:- 1.1 CONNECT 403 TCP_DENIED:HIER_NONE api.github.com:8443 "curl/7.81.0"\n'
-      );
-
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait with non-zero exit code
-      mockExecaFn.mockResolvedValueOnce({ stdout: '1', stderr: '', exitCode: 0 } as any);
-
-      const result = await runAgentCommand(testDir, ['github.com']);
-
-      expect(result.exitCode).toBe(1);
-      // api.github.com should be blocked because port 8443 is not allowed
-      expect(result.blockedDomains).toContain('api.github.com');
-    });
-
-    it('should return empty blockedDomains when no access log exists', async () => {
-      // Don't create access.log
-
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait
-      mockExecaFn.mockResolvedValueOnce({ stdout: '0', stderr: '', exitCode: 0 } as any);
-
-      const result = await runAgentCommand(testDir, ['github.com']);
-
-      expect(result.exitCode).toBe(0);
-      expect(result.blockedDomains).toEqual([]);
-    });
-
-    it('should return exit code 124 when agent times out', async () => {
-      jest.useFakeTimers();
-
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait - never resolves (simulates long-running command)
-      mockExecaFn.mockReturnValueOnce(new Promise(() => {}));
-      // Mock docker stop
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-
-      const resultPromise = runAgentCommand(testDir, ['github.com'], undefined, 1);
-
-      // Use advanceTimersByTimeAsync to flush microtasks between timer advances
-      // This handles the 60s timeout AND the subsequent 500ms log flush delay
-      await jest.advanceTimersByTimeAsync(60 * 1000 + 1000);
-
-      const result = await resultPromise;
-
-      expect(result.exitCode).toBe(124);
-      // Verify docker stop was called
-      expect(mockExecaFn).toHaveBeenCalledWith('docker', ['stop', '-t', '10', 'awf-agent'], expect.objectContaining({ reject: false }));
-
-      jest.useRealTimers();
-    });
-
-    it('should return normal exit code when agent completes before timeout', async () => {
-      jest.useFakeTimers();
-
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait - resolves immediately with exit code 0
-      mockExecaFn.mockResolvedValueOnce({ stdout: '0', stderr: '', exitCode: 0 } as any);
-
-      const resultPromise = runAgentCommand(testDir, ['github.com'], undefined, 30);
-
-      // Advance past the 500ms log flush delay
-      await jest.advanceTimersByTimeAsync(1000);
-
-      const result = await resultPromise;
-
-      expect(result.exitCode).toBe(0);
-      expect(result.blockedDomains).toEqual([]);
-
-      jest.useRealTimers();
-    });
-
-    it('should skip post-run analysis when agent was externally killed', async () => {
-      // Create access.log with denied entries — these should be ignored
-      const squidLogsDir = path.join(testDir, 'squid-logs');
-      fs.mkdirSync(squidLogsDir, { recursive: true });
-      fs.writeFileSync(
-        path.join(squidLogsDir, 'access.log'),
-        '1760994429.358 172.30.0.20:36274 blocked.com:443 -:- 1.1 CONNECT 403 TCP_DENIED:HIER_NONE blocked.com:443 "curl/7.81.0"\n'
-      );
-
-      // Simulate fastKillAgentContainer having been called
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any); // fastKill docker stop
-      await fastKillAgentContainer();
-
-      // Mock docker logs -f
-      mockExecaFn.mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 } as any);
-      // Mock docker wait — container was stopped externally, returns 143
-      mockExecaFn.mockResolvedValueOnce({ stdout: '143', stderr: '', exitCode: 0 } as any);
-
-      const result = await runAgentCommand(testDir, ['github.com']);
-
-      // Should return 143 and skip log analysis (empty blockedDomains)
-      expect(result.exitCode).toBe(143);
-      expect(result.blockedDomains).toEqual([]);
-    });
-  });
-
-  describe('cleanup', () => {
-    let testDir: string;
-
-    beforeEach(() => {
-      testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-'));
-      jest.clearAllMocks();
-      // Mock execa.sync for chmod
-      mockExecaSync.mockReturnValue({ stdout: '', stderr: '', exitCode: 0 });
-    });
-
-    afterEach(() => {
-      // Clean up any remaining test directories
-      if (fs.existsSync(testDir)) {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-      // Clean up any moved log directories
-      const timestamp = path.basename(testDir).replace('awf-', '');
-      const agentLogsDir = path.join(os.tmpdir(), `awf-agent-logs-${timestamp}`);
-      const squidLogsDir = path.join(os.tmpdir(), `squid-logs-${timestamp}`);
-      if (fs.existsSync(agentLogsDir)) {
-        fs.rmSync(agentLogsDir, { recursive: true, force: true });
-      }
-      if (fs.existsSync(squidLogsDir)) {
-        fs.rmSync(squidLogsDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should skip cleanup when keepFiles is true', async () => {
-      await cleanup(testDir, true);
-
-      // Verify directory still exists
-      expect(fs.existsSync(testDir)).toBe(true);
-    });
-
-    it('should remove work directory when keepFiles is false', async () => {
-      await cleanup(testDir, false);
-
-      expect(fs.existsSync(testDir)).toBe(false);
-    });
-
-    it('should clean up chroot-home directory alongside workDir', async () => {
-      // Create chroot-home sibling directory (as writeConfigs does in chroot mode)
-      const chrootHomeDir = `${testDir}-chroot-home`;
-      fs.mkdirSync(chrootHomeDir, { recursive: true });
-
-      await cleanup(testDir, false);
-
-      // Both workDir and chroot-home should be removed
-      expect(fs.existsSync(testDir)).toBe(false);
-      expect(fs.existsSync(chrootHomeDir)).toBe(false);
-    });
-
-    it('should preserve agent logs when they exist', async () => {
-      // Create agent logs directory with a file
-      const agentLogsDir = path.join(testDir, 'agent-logs');
-      fs.mkdirSync(agentLogsDir, { recursive: true });
-      fs.writeFileSync(path.join(agentLogsDir, 'test.log'), 'test log content');
-
-      await cleanup(testDir, false);
-
-      // Verify work directory was removed
-      expect(fs.existsSync(testDir)).toBe(false);
-
-      // Verify agent logs were moved
-      const timestamp = path.basename(testDir).replace('awf-', '');
-      const preservedLogsDir = path.join(os.tmpdir(), `awf-agent-logs-${timestamp}`);
-      expect(fs.existsSync(preservedLogsDir)).toBe(true);
-      expect(fs.readFileSync(path.join(preservedLogsDir, 'test.log'), 'utf-8')).toBe('test log content');
-    });
-
-    it('should preserve squid logs when they exist', async () => {
-      // Create squid logs directory with a file
-      const squidLogsDir = path.join(testDir, 'squid-logs');
-      fs.mkdirSync(squidLogsDir, { recursive: true });
-      fs.writeFileSync(path.join(squidLogsDir, 'access.log'), 'squid log content');
-
-      await cleanup(testDir, false);
-
-      // Verify work directory was removed
-      expect(fs.existsSync(testDir)).toBe(false);
-
-      // Verify squid logs were moved
-      const timestamp = path.basename(testDir).replace('awf-', '');
-      const preservedLogsDir = path.join(os.tmpdir(), `squid-logs-${timestamp}`);
-      expect(fs.existsSync(preservedLogsDir)).toBe(true);
-    });
-
-    it('should not preserve empty log directories', async () => {
-      // Create empty agent logs directory
-      const agentLogsDir = path.join(testDir, 'agent-logs');
-      fs.mkdirSync(agentLogsDir, { recursive: true });
-
-      await cleanup(testDir, false);
-
-      // Verify work directory was removed
-      expect(fs.existsSync(testDir)).toBe(false);
-
-      // Verify no empty log directory was created
-      const timestamp = path.basename(testDir).replace('awf-', '');
-      const preservedLogsDir = path.join(os.tmpdir(), `awf-agent-logs-${timestamp}`);
-      expect(fs.existsSync(preservedLogsDir)).toBe(false);
-    });
-
-    it('should use proxyLogsDir when specified', async () => {
-      const proxyLogsDir = path.join(testDir, 'custom-proxy-logs');
-      fs.mkdirSync(proxyLogsDir, { recursive: true });
-      fs.writeFileSync(path.join(proxyLogsDir, 'access.log'), 'proxy log content');
-
-      await cleanup(testDir, false, proxyLogsDir);
-
-      // Verify chmod was called on proxyLogsDir
-      expect(mockExecaSync).toHaveBeenCalledWith('chmod', ['-R', 'a+rX', proxyLogsDir]);
-    });
-
-    it('should not move squid logs to /tmp when proxyLogsDir is specified', async () => {
-      // proxyLogsDir must be OUTSIDE workDir since cleanup deletes workDir
-      const externalDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-proxy-logs-test-'));
-      const proxyLogsDir = path.join(externalDir, 'proxy-logs');
-      fs.mkdirSync(proxyLogsDir, { recursive: true });
-      fs.writeFileSync(path.join(proxyLogsDir, 'access.log'), 'proxy log content');
-
-      try {
-        await cleanup(testDir, false, proxyLogsDir);
-
-        // Logs should remain in proxyLogsDir (not moved to /tmp/squid-logs-*)
-        expect(fs.existsSync(path.join(proxyLogsDir, 'access.log'))).toBe(true);
-      } finally {
-        fs.rmSync(externalDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should chmod api-proxy-logs subdirectory when proxyLogsDir is specified', async () => {
-      // proxyLogsDir must be OUTSIDE workDir since cleanup deletes workDir
-      const externalDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-proxy-logs-test-'));
-      const proxyLogsDir = path.join(externalDir, 'proxy-logs');
-      const apiProxyLogsDir = path.join(proxyLogsDir, 'api-proxy-logs');
-      fs.mkdirSync(proxyLogsDir, { recursive: true });
-      fs.mkdirSync(apiProxyLogsDir, { recursive: true });
-      fs.writeFileSync(path.join(proxyLogsDir, 'access.log'), 'proxy log content');
-      fs.writeFileSync(path.join(apiProxyLogsDir, 'access.log'), 'api proxy log content');
-
-      try {
-        await cleanup(testDir, false, proxyLogsDir);
-
-        // Verify chmod was called on both proxyLogsDir and api-proxy-logs subdirectory
-        expect(mockExecaSync).toHaveBeenCalledWith('chmod', ['-R', 'a+rX', proxyLogsDir]);
-        expect(mockExecaSync).toHaveBeenCalledWith('chmod', ['-R', 'a+rX', apiProxyLogsDir]);
-      } finally {
-        fs.rmSync(externalDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should handle non-existent work directory gracefully', async () => {
-      const nonExistentDir = path.join(os.tmpdir(), 'awf-nonexistent-12345');
-
-      // Should not throw
-      await expect(cleanup(nonExistentDir, false)).resolves.not.toThrow();
-    });
-
-    it('should preserve session state to /tmp when sessionStateDir is not specified', async () => {
-      const sessionStateDir = path.join(testDir, 'agent-session-state');
-      const sessionDir = path.join(sessionStateDir, 'abc-123');
-      fs.mkdirSync(sessionDir, { recursive: true });
-      fs.writeFileSync(path.join(sessionDir, 'events.jsonl'), '{"event":"test"}');
-
-      await cleanup(testDir, false);
-
-      // Verify session state was moved to timestamped /tmp directory
-      const timestamp = path.basename(testDir).replace('awf-', '');
-      const preservedDir = path.join(os.tmpdir(), `awf-agent-session-state-${timestamp}`);
-      expect(fs.existsSync(preservedDir)).toBe(true);
-      expect(fs.existsSync(path.join(preservedDir, 'abc-123', 'events.jsonl'))).toBe(true);
-    });
-
-    it('should chmod session state in-place when sessionStateDir is specified', async () => {
-      const externalDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-session-test-'));
-      const sessionStateDir = path.join(externalDir, 'session-state');
-      fs.mkdirSync(sessionStateDir, { recursive: true });
-      fs.writeFileSync(path.join(sessionStateDir, 'events.jsonl'), '{"event":"test"}');
-
-      try {
-        await cleanup(testDir, false, undefined, undefined, sessionStateDir);
-
-        // Verify chmod was called on sessionStateDir (not moved)
-        expect(mockExecaSync).toHaveBeenCalledWith('chmod', ['-R', 'a+rX', sessionStateDir]);
-        // Files should remain in-place
-        expect(fs.existsSync(path.join(sessionStateDir, 'events.jsonl'))).toBe(true);
-      } finally {
-        fs.rmSync(externalDir, { recursive: true, force: true });
-      }
-    });
-  });
-
-  describe('readGitHubPathEntries', () => {
-    let tmpDir: string;
-
-    beforeEach(() => {
-      tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-github-path-'));
-    });
-
-    afterEach(() => {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    });
-
-    it('should return empty array when GITHUB_PATH is not set', () => {
-      const originalGithubPath = process.env.GITHUB_PATH;
-      delete process.env.GITHUB_PATH;
-
-      try {
-        const result = readGitHubPathEntries();
-        expect(result).toEqual([]);
-      } finally {
-        if (originalGithubPath !== undefined) {
-          process.env.GITHUB_PATH = originalGithubPath;
-        }
-      }
-    });
-
-    it('should return empty array when GITHUB_PATH file does not exist', () => {
-      const originalGithubPath = process.env.GITHUB_PATH;
-      process.env.GITHUB_PATH = '/nonexistent/path/to/github_path_file';
-
-      try {
-        const result = readGitHubPathEntries();
-        expect(result).toEqual([]);
-      } finally {
-        if (originalGithubPath !== undefined) {
-          process.env.GITHUB_PATH = originalGithubPath;
-        } else {
-          delete process.env.GITHUB_PATH;
-        }
-      }
-    });
-
-    it('should read path entries from GITHUB_PATH file', () => {
-      const pathFile = path.join(tmpDir, 'add_path');
-      fs.writeFileSync(pathFile, '/opt/hostedtoolcache/Ruby/3.3.10/x64/bin\n/opt/hostedtoolcache/Python/3.12.0/x64/bin\n');
-
-      const originalGithubPath = process.env.GITHUB_PATH;
-      process.env.GITHUB_PATH = pathFile;
-
-      try {
-        const result = readGitHubPathEntries();
-        expect(result).toEqual([
-          '/opt/hostedtoolcache/Ruby/3.3.10/x64/bin',
-          '/opt/hostedtoolcache/Python/3.12.0/x64/bin',
-        ]);
-      } finally {
-        if (originalGithubPath !== undefined) {
-          process.env.GITHUB_PATH = originalGithubPath;
-        } else {
-          delete process.env.GITHUB_PATH;
-        }
-      }
-    });
-
-    it('should handle empty lines and whitespace in GITHUB_PATH file', () => {
-      const pathFile = path.join(tmpDir, 'add_path');
-      fs.writeFileSync(pathFile, '  /opt/hostedtoolcache/Ruby/3.3.10/x64/bin  \n\n  \n/opt/dart-sdk/bin\n');
-
-      const originalGithubPath = process.env.GITHUB_PATH;
-      process.env.GITHUB_PATH = pathFile;
-
-      try {
-        const result = readGitHubPathEntries();
-        expect(result).toEqual([
-          '/opt/hostedtoolcache/Ruby/3.3.10/x64/bin',
-          '/opt/dart-sdk/bin',
-        ]);
-      } finally {
-        if (originalGithubPath !== undefined) {
-          process.env.GITHUB_PATH = originalGithubPath;
-        } else {
-          delete process.env.GITHUB_PATH;
-        }
-      }
-    });
-
-    it('should handle empty GITHUB_PATH file', () => {
-      const pathFile = path.join(tmpDir, 'add_path');
-      fs.writeFileSync(pathFile, '');
-
-      const originalGithubPath = process.env.GITHUB_PATH;
-      process.env.GITHUB_PATH = pathFile;
-
-      try {
-        const result = readGitHubPathEntries();
-        expect(result).toEqual([]);
-      } finally {
-        if (originalGithubPath !== undefined) {
-          process.env.GITHUB_PATH = originalGithubPath;
-        } else {
-          delete process.env.GITHUB_PATH;
-        }
-      }
-    });
-  });
-
-  describe('parseGitHubEnvFile', () => {
-    it('should parse simple KEY=VALUE entries', () => {
-      const result = parseGitHubEnvFile('GOROOT=/usr/local/go\nJAVA_HOME=/usr/lib/jvm/java-17\n');
-      expect(result).toEqual({
-        GOROOT: '/usr/local/go',
-        JAVA_HOME: '/usr/lib/jvm/java-17',
-      });
-    });
-
-    it('should handle values containing = characters', () => {
-      const result = parseGitHubEnvFile('MY_VAR=key=value=extra\n');
-      expect(result).toEqual({ MY_VAR: 'key=value=extra' });
-    });
-
-    it('should handle heredoc multiline values', () => {
-      const content = 'MULTI_LINE<<EOF\nline1\nline2\nline3\nEOF\n';
-      const result = parseGitHubEnvFile(content);
-      expect(result).toEqual({ MULTI_LINE: 'line1\nline2\nline3' });
-    });
-
-    it('should handle CRLF line endings', () => {
-      const result = parseGitHubEnvFile('GOROOT=/usr/local/go\r\nJAVA_HOME=/usr/lib/jvm\r\n');
-      expect(result).toEqual({
-        GOROOT: '/usr/local/go',
-        JAVA_HOME: '/usr/lib/jvm',
-      });
-    });
-
-    it('should handle mixed simple and heredoc entries', () => {
-      const content = 'SIMPLE=value\nHEREDOC<<END\nmulti\nline\nEND\nANOTHER=val2\n';
-      const result = parseGitHubEnvFile(content);
-      expect(result).toEqual({
-        SIMPLE: 'value',
-        HEREDOC: 'multi\nline',
-        ANOTHER: 'val2',
-      });
-    });
-
-    it('should skip empty lines', () => {
-      const result = parseGitHubEnvFile('\n\nGOROOT=/go\n\n');
-      expect(result).toEqual({ GOROOT: '/go' });
-    });
-
-    it('should return empty object for empty content', () => {
-      expect(parseGitHubEnvFile('')).toEqual({});
-    });
-
-    it('should handle unterminated heredoc gracefully', () => {
-      const content = 'BROKEN<<EOF\nline1\nline2';
-      const result = parseGitHubEnvFile(content);
-      expect(result).toEqual({ BROKEN: 'line1\nline2' });
-    });
-  });
-
-  describe('readGitHubEnvEntries', () => {
-    let tmpDir: string;
-
-    beforeEach(() => {
-      tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-github-env-'));
-    });
-
-    afterEach(() => {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    });
-
-    it('should return empty object when GITHUB_ENV is not set', () => {
-      const original = process.env.GITHUB_ENV;
-      delete process.env.GITHUB_ENV;
-
-      try {
-        const result = readGitHubEnvEntries();
-        expect(result).toEqual({});
-      } finally {
-        if (original !== undefined) process.env.GITHUB_ENV = original;
-        else delete process.env.GITHUB_ENV;
-      }
-    });
-
-    it('should read entries from GITHUB_ENV file', () => {
-      const original = process.env.GITHUB_ENV;
-      const envFile = path.join(tmpDir, 'github_env');
-      fs.writeFileSync(envFile, 'GOROOT=/usr/local/go\nCARGO_HOME=/home/.cargo\n');
-      process.env.GITHUB_ENV = envFile;
-
-      try {
-        const result = readGitHubEnvEntries();
-        expect(result.GOROOT).toBe('/usr/local/go');
-        expect(result.CARGO_HOME).toBe('/home/.cargo');
-      } finally {
-        if (original !== undefined) process.env.GITHUB_ENV = original;
-        else delete process.env.GITHUB_ENV;
-      }
-    });
-
-    it('should return empty object when file does not exist', () => {
-      const original = process.env.GITHUB_ENV;
-      process.env.GITHUB_ENV = '/nonexistent/path/github_env';
-
-      try {
-        const result = readGitHubEnvEntries();
-        expect(result).toEqual({});
-      } finally {
-        if (original !== undefined) process.env.GITHUB_ENV = original;
-        else delete process.env.GITHUB_ENV;
-      }
-    });
-  });
-
   describe('toolchain var fallback to GITHUB_ENV', () => {
     let tmpDir: string;
     const testConfig: WrapperConfig = {
@@ -5041,53 +3407,6 @@ describe('docker-manager', () => {
         if (savedGithubEnv !== undefined) process.env.GITHUB_ENV = savedGithubEnv;
         else delete process.env.GITHUB_ENV;
       }
-    });
-  });
-
-  describe('mergeGitHubPathEntries', () => {
-    it('should return current PATH when no github path entries', () => {
-      const result = mergeGitHubPathEntries('/usr/bin:/usr/local/bin', []);
-      expect(result).toBe('/usr/bin:/usr/local/bin');
-    });
-
-    it('should prepend github path entries to current PATH', () => {
-      const result = mergeGitHubPathEntries(
-        '/usr/bin:/usr/local/bin',
-        ['/opt/hostedtoolcache/Ruby/3.3.10/x64/bin']
-      );
-      expect(result).toBe('/opt/hostedtoolcache/Ruby/3.3.10/x64/bin:/usr/bin:/usr/local/bin');
-    });
-
-    it('should not duplicate entries already in PATH', () => {
-      const result = mergeGitHubPathEntries(
-        '/opt/hostedtoolcache/Ruby/3.3.10/x64/bin:/usr/bin:/usr/local/bin',
-        ['/opt/hostedtoolcache/Ruby/3.3.10/x64/bin']
-      );
-      expect(result).toBe('/opt/hostedtoolcache/Ruby/3.3.10/x64/bin:/usr/bin:/usr/local/bin');
-    });
-
-    it('should handle multiple new entries', () => {
-      const result = mergeGitHubPathEntries(
-        '/usr/bin',
-        ['/opt/hostedtoolcache/Ruby/3.3.10/x64/bin', '/opt/dart-sdk/bin']
-      );
-      expect(result).toBe('/opt/hostedtoolcache/Ruby/3.3.10/x64/bin:/opt/dart-sdk/bin:/usr/bin');
-    });
-
-    it('should handle mix of new and existing entries', () => {
-      const result = mergeGitHubPathEntries(
-        '/opt/hostedtoolcache/Ruby/3.3.10/x64/bin:/usr/bin',
-        ['/opt/hostedtoolcache/Ruby/3.3.10/x64/bin', '/opt/dart-sdk/bin']
-      );
-      expect(result).toBe('/opt/dart-sdk/bin:/opt/hostedtoolcache/Ruby/3.3.10/x64/bin:/usr/bin');
-    });
-
-    it('should handle empty current PATH', () => {
-      const result = mergeGitHubPathEntries(
-        '',
-        ['/opt/hostedtoolcache/Ruby/3.3.10/x64/bin']
-      );
-      expect(result).toBe('/opt/hostedtoolcache/Ruby/3.3.10/x64/bin');
     });
   });
 
@@ -5203,270 +3522,4 @@ describe('docker-manager', () => {
     });
   });
 
-  describe('readEnvFile', () => {
-    let tmpDir: string;
-
-    beforeEach(() => {
-      tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-test-readenvfile-'));
-    });
-
-    afterEach(() => {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    });
-
-    it('should parse KEY=VALUE pairs', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, 'FOO=bar\nBAZ=qux\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: 'bar', BAZ: 'qux' });
-    });
-
-    it('should skip comment lines starting with #', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, '# comment\nFOO=bar\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: 'bar' });
-    });
-
-    it('should skip blank lines', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, '\nFOO=bar\n\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: 'bar' });
-    });
-
-    it('should allow empty values', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, 'FOO=\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: '' });
-    });
-
-    it('should allow values containing = signs', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, 'FOO=a=b=c\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: 'a=b=c' });
-    });
-
-    it('should ignore lines that do not match KEY=VALUE format', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, 'INVALID LINE\nFOO=bar\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: 'bar' });
-    });
-
-    it('should reject keys starting with a digit', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, '123KEY=value\nFOO=bar\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: 'bar' });
-    });
-
-    it('should reject keys containing hyphens', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, 'KEY-NAME=value\nFOO=bar\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: 'bar' });
-    });
-
-    it('should handle lines with leading whitespace by trimming them', () => {
-      const envFile = path.join(tmpDir, '.env');
-      fs.writeFileSync(envFile, '  FOO=bar\n');
-      expect(readEnvFile(envFile)).toEqual({ FOO: 'bar' });
-    });
-
-    it('should throw when file does not exist', () => {
-      expect(() => readEnvFile(path.join(tmpDir, 'missing.env'))).toThrow();
-    });
-  });
-
-  describe('collectDiagnosticLogs', () => {
-    let testDir: string;
-
-    beforeEach(() => {
-      testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-'));
-      jest.clearAllMocks();
-    });
-
-    afterEach(() => {
-      if (fs.existsSync(testDir)) {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should create diagnostics directory and write container logs', async () => {
-      // Mock docker logs returning content
-      mockExecaFn
-        .mockResolvedValueOnce({ stdout: 'squid log output', stderr: '', exitCode: 0 })  // docker logs awf-squid
-        .mockResolvedValueOnce({ stdout: '0 ', stderr: '', exitCode: 0 })                 // docker inspect state awf-squid
-        .mockResolvedValueOnce({ stdout: '[{"Type":"bind"}]', stderr: '', exitCode: 0 })  // docker inspect mounts awf-squid
-        .mockResolvedValueOnce({ stdout: 'agent log output', stderr: '', exitCode: 0 })   // docker logs awf-agent
-        .mockResolvedValueOnce({ stdout: '1 container crashed', stderr: '', exitCode: 0 }) // docker inspect state awf-agent
-        .mockResolvedValueOnce({ stdout: '[{"Type":"volume"}]', stderr: '', exitCode: 0 }) // docker inspect mounts awf-agent
-        .mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 1 })                    // docker logs awf-api-proxy (not started)
-        .mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 1 })                    // docker inspect state awf-api-proxy
-        .mockResolvedValueOnce({ stdout: 'null', stderr: '', exitCode: 0 })                // docker inspect mounts awf-api-proxy (null)
-        .mockResolvedValueOnce({ stdout: '', stderr: '', exitCode: 0 })                    // docker logs awf-iptables-init
-        .mockResolvedValueOnce({ stdout: '0 ', stderr: '', exitCode: 0 })                 // docker inspect state awf-iptables-init
-        .mockResolvedValueOnce({ stdout: '[]', stderr: '', exitCode: 0 });                 // docker inspect mounts awf-iptables-init
-
-      // Create a docker-compose.yml with a secret env var
-      const composeContent = [
-        'services:',
-        '  squid:',
-        '    environment:',
-        '      AWF_SQUID_CONFIG_B64: secretvalue',
-        '      GITHUB_TOKEN: ghp_abc123',
-        '      SOME_KEY: mykey',
-        '      NORMAL_VAR: normalvalue',
-      ].join('\n');
-      fs.writeFileSync(path.join(testDir, 'docker-compose.yml'), composeContent);
-
-      await collectDiagnosticLogs(testDir);
-
-      const diagnosticsDir = path.join(testDir, 'diagnostics');
-      expect(fs.existsSync(diagnosticsDir)).toBe(true);
-
-      // awf-squid.log should have content
-      expect(fs.existsSync(path.join(diagnosticsDir, 'awf-squid.log'))).toBe(true);
-      expect(fs.readFileSync(path.join(diagnosticsDir, 'awf-squid.log'), 'utf8')).toContain('squid log output');
-
-      // awf-agent.log should have content
-      expect(fs.existsSync(path.join(diagnosticsDir, 'awf-agent.log'))).toBe(true);
-      expect(fs.readFileSync(path.join(diagnosticsDir, 'awf-agent.log'), 'utf8')).toContain('agent log output');
-
-      // State files should be written
-      expect(fs.existsSync(path.join(diagnosticsDir, 'awf-squid.state'))).toBe(true);
-      expect(fs.existsSync(path.join(diagnosticsDir, 'awf-agent.state'))).toBe(true);
-
-      // Mounts files for containers that returned non-null JSON
-      expect(fs.existsSync(path.join(diagnosticsDir, 'awf-squid.mounts.json'))).toBe(true);
-      expect(fs.existsSync(path.join(diagnosticsDir, 'awf-agent.mounts.json'))).toBe(true);
-      // awf-api-proxy returned 'null' so no mounts file
-      expect(fs.existsSync(path.join(diagnosticsDir, 'awf-api-proxy.mounts.json'))).toBe(false);
-
-      // Sanitized docker-compose.yml should exist with secrets redacted
-      const sanitizedCompose = fs.readFileSync(path.join(diagnosticsDir, 'docker-compose.yml'), 'utf8');
-      expect(sanitizedCompose).toContain('[REDACTED]');
-      // GITHUB_TOKEN and SOME_KEY contain TOKEN/KEY → redacted
-      expect(sanitizedCompose).not.toContain('ghp_abc123');
-      expect(sanitizedCompose).not.toContain('mykey');
-      // AWF_SQUID_CONFIG_B64 does not contain TOKEN/KEY/SECRET → preserved
-      expect(sanitizedCompose).toContain('secretvalue');
-      // Non-secret env vars should be unchanged
-      expect(sanitizedCompose).toContain('NORMAL_VAR: normalvalue');
-    });
-
-    it('should not write log file when docker logs returns empty output', async () => {
-      mockExecaFn
-        .mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 }); // all containers return empty
-
-      await collectDiagnosticLogs(testDir);
-
-      const diagnosticsDir = path.join(testDir, 'diagnostics');
-      // No .log files should be written for empty output
-      const files = fs.existsSync(diagnosticsDir) ? fs.readdirSync(diagnosticsDir) : [];
-      const logFiles = files.filter(f => f.endsWith('.log'));
-      expect(logFiles).toHaveLength(0);
-    });
-
-    it('should skip docker-compose.yml sanitization when file does not exist', async () => {
-      mockExecaFn.mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 });
-
-      // No docker-compose.yml created in testDir
-      await expect(collectDiagnosticLogs(testDir)).resolves.not.toThrow();
-
-      const diagnosticsDir = path.join(testDir, 'diagnostics');
-      expect(fs.existsSync(path.join(diagnosticsDir, 'docker-compose.yml'))).toBe(false);
-    });
-
-    it('should handle docker command failures gracefully', async () => {
-      // All docker commands throw errors
-      mockExecaFn.mockRejectedValue(new Error('docker not found'));
-
-      await expect(collectDiagnosticLogs(testDir)).resolves.not.toThrow();
-    });
-
-    it('should redact lowercase and mixed-case secret env var names', async () => {
-      mockExecaFn.mockResolvedValue({ stdout: '', stderr: '', exitCode: 0 });
-
-      const composeContent = [
-        'services:',
-        '  agent:',
-        '    environment:',
-        '      github_token: ghp_lowercase',
-        '      Api_Key: mixedcase_value',
-        '      OAUTH_SECRET: uppercase_secret',
-        '      not_sensitive: keepme',
-      ].join('\n');
-      fs.writeFileSync(path.join(testDir, 'docker-compose.yml'), composeContent);
-
-      await collectDiagnosticLogs(testDir);
-
-      const sanitized = fs.readFileSync(path.join(testDir, 'diagnostics', 'docker-compose.yml'), 'utf8');
-      // All three secret patterns should be redacted
-      expect(sanitized).not.toContain('ghp_lowercase');
-      expect(sanitized).not.toContain('mixedcase_value');
-      expect(sanitized).not.toContain('uppercase_secret');
-      // Non-secret var must be preserved
-      expect(sanitized).toContain('not_sensitive: keepme');
-    });
-  });
-
-  describe('cleanup - diagnostics preservation', () => {
-    let testDir: string;
-
-    beforeEach(() => {
-      testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-'));
-      jest.clearAllMocks();
-      mockExecaSync.mockReturnValue({ stdout: '', stderr: '', exitCode: 0 });
-    });
-
-    afterEach(() => {
-      if (fs.existsSync(testDir)) {
-        fs.rmSync(testDir, { recursive: true, force: true });
-      }
-      const timestamp = path.basename(testDir).replace('awf-', '');
-      const diagDir = path.join(os.tmpdir(), `awf-diagnostics-${timestamp}`);
-      if (fs.existsSync(diagDir)) {
-        fs.rmSync(diagDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should preserve diagnostics to /tmp when no auditDir is specified', async () => {
-      const diagnosticsDir = path.join(testDir, 'diagnostics');
-      fs.mkdirSync(diagnosticsDir, { recursive: true });
-      fs.writeFileSync(path.join(diagnosticsDir, 'awf-squid.log'), 'squid crashed\n');
-
-      await cleanup(testDir, false);
-
-      const timestamp = path.basename(testDir).replace('awf-', '');
-      const preserved = path.join(os.tmpdir(), `awf-diagnostics-${timestamp}`);
-      expect(fs.existsSync(preserved)).toBe(true);
-      expect(fs.readFileSync(path.join(preserved, 'awf-squid.log'), 'utf8')).toBe('squid crashed\n');
-    });
-
-    it('should co-locate diagnostics under auditDir/diagnostics when auditDir is specified', async () => {
-      const auditDir = fs.mkdtempSync(path.join(os.tmpdir(), 'awf-audit-test-'));
-      try {
-        const diagnosticsDir = path.join(testDir, 'diagnostics');
-        fs.mkdirSync(diagnosticsDir, { recursive: true });
-        fs.writeFileSync(path.join(diagnosticsDir, 'awf-agent.log'), 'agent output\n');
-
-        await cleanup(testDir, false, undefined, auditDir);
-
-        const auditDiagnosticsDir = path.join(auditDir, 'diagnostics');
-        expect(fs.existsSync(auditDiagnosticsDir)).toBe(true);
-        expect(fs.readFileSync(path.join(auditDiagnosticsDir, 'awf-agent.log'), 'utf8')).toBe('agent output\n');
-        expect(mockExecaSync).toHaveBeenCalledWith('chmod', ['-R', 'a+rX', auditDiagnosticsDir]);
-      } finally {
-        fs.rmSync(auditDir, { recursive: true, force: true });
-      }
-    });
-
-    it('should not create diagnostics destination when diagnostics dir is empty', async () => {
-      // Empty diagnostics dir
-      const diagnosticsDir = path.join(testDir, 'diagnostics');
-      fs.mkdirSync(diagnosticsDir, { recursive: true });
-
-      await cleanup(testDir, false);
-
-      const timestamp = path.basename(testDir).replace('awf-', '');
-      const preserved = path.join(os.tmpdir(), `awf-diagnostics-${timestamp}`);
-      expect(fs.existsSync(preserved)).toBe(false);
-    });
-  });
 });
